Extract runstate file
